+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 ### Added
 
+* `pathtools.join2` can be used to join paths, much like `os.path.join` except
+  that it will work with `java.io.File` objects as well (but doesn't support
+  more than two path components / parameters).)
+
 ### Changed
 
 * `strtools.filename` and `pathtools.parse_path` can now also work on